+++ CHANGELOG.md
@@ -28,6 +28,7 @@
 24. New code highlight system
 25. New submit code button
 26. Fetch test cases: pbinfo.ro support
+27. Fetch test cases: kilonova.ro support
 
 ### Fixed/Changes: